+++ spec/unit/libraries/CMCredentials_spec.rb
@@ -41,6 +41,10 @@ class LoadedDummyResource
   def monitoring_agent_token
     return 'resource token'
   end
+
+  def monitoring_mock_api
+    return 'resource mocking'
+  end
 end
 
 # An empty class to fake no resource variables
@@ -75,6 +79,8 @@ class TestData
       }
       data['rackspace_cloudmonitoring']['auth']['url'] = 'node authurl'
       data['rackspace_cloudmonitoring']['config'] = { 'agent' => { 'token' => 'node token' } }
+      data['rackspace_cloudmonitoring']['mock'] = 'node mocking'
+      data['rackspace_cloudmonitoring']['api'] = { 'pagination_limit' => 'node pagination_limit' }
     end
 
     return data
@@ -83,10 +89,42 @@ class TestData
   # iteration_data: A hash to be looped over to test the various attributes handled
   def self.iteration_data
     return {
-      api_key:  'apikey',
-      username: 'username',
-      auth_url: 'authurl',
-      token:    'token'
+      api_key: {
+        name: 'apikey',
+        resource_supported: true,
+        node_supported: true,
+        databag_supported: true
+      },
+      username: {
+        name: 'username',
+        resource_supported: true,
+        node_supported: true,
+        databag_supported: true
+      },
+      auth_url: {
+        name: 'authurl',
+        resource_supported: true,
+        node_supported: true,
+        databag_supported: true
+      },
+      token: {
+        name: 'token',
+        resource_supported: true,
+        node_supported: true,
+        databag_supported: true
+      },
+      mocking: {
+        name: 'mocking',
+        resource_supported: true,
+        node_supported: true,
+        databag_supported: false
+      },
+      pagination_limit: {
+        name: 'pagination_limit',
+        resource_supported: false,
+        node_supported: true,
+        databag_supported: false
+      }
     }
   end
 
@@ -147,37 +185,41 @@ describe CMCredentials do
     end
   end
 
-  TestData.iteration_data.each do |key, value|
+  TestData.iteration_data.each do |key, data|
     describe "#_get_resource_attribute(#{key})" do
       before :each do
         TestData.stub_databag(true)
       end
 
-      it 'returns available resource data' do
-        test_creds = CMCredentials.new(nil, LoadedDummyResource.new)
-        test_creds._get_resource_attribute(key).should eql "resource #{value}"
+      if data[:resource_supported]
+        it 'returns available resource data' do
+          test_creds = CMCredentials.new(nil, LoadedDummyResource.new)
+          test_creds._get_resource_attribute(key).should eql "resource #{data[:name]}"
+        end
       end
 
       it 'returns nil for unavailable data' do
         test_creds = CMCredentials.new(nil, EmptyDummyResource.new)
         test_creds._get_resource_attribute(key).should eql nil
       end
-    end
 
-    describe "#_get_node_attribute(#{key})" do
-      before :each do
-        TestData.stub_databag(true)
-        @test_node_data = TestData.node_data
-      end
+      describe "#_get_node_attribute(#{key})" do
+        before :each do
+          TestData.stub_databag(true)
+          @test_node_data = TestData.node_data
+        end
 
-      it 'returns available node data' do
-        test_creds = CMCredentials.new(@test_node_data, LoadedDummyResource.new)
-        test_creds._get_node_attribute(key).should eql "node #{value}"
-      end
+        if data[:node_supported]
+          it 'returns available node data' do
+            test_creds = CMCredentials.new(@test_node_data, LoadedDummyResource.new)
+            test_creds._get_node_attribute(key).should eql "node #{data[:name]}"
+          end
+        end
 
-      it 'returns nil for unavailable data' do
-        test_creds = CMCredentials.new(nil, EmptyDummyResource.new)
-        test_creds._get_node_attribute(key).should eql nil
+        it 'returns nil for unavailable data' do
+          test_creds = CMCredentials.new(nil, EmptyDummyResource.new)
+          test_creds._get_node_attribute(key).should eql nil
+        end
       end
     end
 
@@ -192,10 +234,12 @@ describe CMCredentials do
         test_creds._get_databag_attribute(key).should eql nil
       end
 
-      it 'returns available databag data' do
-        TestData.stub_databag(true)
-        test_creds = CMCredentials.new(@test_node_data, LoadedDummyResource.new)
-        test_creds._get_databag_attribute(key).should eql "databag #{value}"
+      if data[:databag_supported]
+        it 'returns available databag data' do
+          TestData.stub_databag(true)
+          test_creds = CMCredentials.new(@test_node_data, LoadedDummyResource.new)
+          test_creds._get_databag_attribute(key).should eql "databag #{data[:name]}"
+        end
       end
 
       it 'returns nil for unavailable data' do
@@ -218,24 +262,30 @@ describe CMCredentials do
         @test_node_data.should be_an_instance_of Hash
       end
 
-      it 'favors resource attributes over all others' do
-        test_creds = CMCredentials.new(@test_node_data, LoadedDummyResource.new)
-        test_creds.get_attribute(key).should eql "resource #{value}"
+      if data[:resource_supported]
+        it 'favors resource attributes over all others' do
+          test_creds = CMCredentials.new(@test_node_data, LoadedDummyResource.new)
+          test_creds.get_attribute(key).should eql "resource #{data[:name]}"
+        end
       end
 
-      it 'favors node attributes over databag with an empty resource' do
-        test_creds = CMCredentials.new(@test_node_data, EmptyDummyResource.new)
-        test_creds.get_attribute(key).should eql "node #{value}"
+      if data[:node_supported]
+        it 'favors node attributes over databag with an empty resource' do
+          test_creds = CMCredentials.new(@test_node_data, EmptyDummyResource.new)
+          test_creds.get_attribute(key).should eql "node #{data[:name]}"
+        end
+
+        it 'favors node attributes over databag with a nil resource' do
+          test_creds = CMCredentials.new(@test_node_data, nil)
+          test_creds.get_attribute(key).should eql "node #{data[:name]}"
+        end
       end
 
-      it 'favors node attributes over databag with a nil resource' do
-        test_creds = CMCredentials.new(@test_node_data, nil)
-        test_creds.get_attribute(key).should eql "node #{value}"
-      end
-
-      it 'Returns databag attributes with no node or resource values' do
-        test_creds = CMCredentials.new(TestData.node_data(true), nil)
-        test_creds.get_attribute(key).should eql "databag #{value}"
+      if data[:databag_supported]
+        it 'Returns databag attributes with no node or resource valuess' do
+          test_creds = CMCredentials.new(TestData.node_data(true), nil)
+          test_creds.get_attribute(key).should eql "databag #{data[:name]}"
+        end
       end
 
       it 'Returns nil with nothing available' do