+++ spec/unit/libraries/CMObjBase_spec.rb
@@ -17,109 +17,55 @@
 
 require 'spec_helper'
 require_relative '../../../libraries/CMObjBase.rb'
+require_relative '../../../libraries/mock_data.rb'
 include Opscode::Rackspace::Monitoring
 
-#
-# Dummy object for testing
-# Behaves loosly like Fog objects
-# Fog is currently mocked, but the mocking has no backing storage so it just returns
-#  random data.  This class is intended to fill the gap.  It is ONLY intended for
-#  testing CMObjBase, it will not have sufficient coverage for the inheriting classes
-#  and LWRPs.
-#
-class DummyObject
-  attr_accessor :id, :label, :saved, :destroyed
-
-  def initialize(arg_hash)
-    @id = arg_hash[:id]
-    @label = arg_hash[:label]
-    @saved = false
-    @destroyed = false
+# CMObjBaseTestHelpers: Helper methods/classes for the tests
+module CMObjBaseTestHelpers
+  # dummy_parent_object: Return a dummy patent object for testing
+  # Code deduplication method: Provides a single definition for what is mocking the object.
+  def dummy_parent_object
+    return MockData::MockMonitoringParent.new(MockData::MockMonitoringEntity)
   end
-
-  #
-  # Methods mimicing Fog
-  #
-  def save
-    @saved = true
-  end
-
-  def destroy
-    @destroyed = true
-  end
-
-  def compare?(other_obj)
-    if @destroyed || other_obj.destroyed
-      return false
-    end
-
-    if @id != other_obj.id
-      return false
-    end
-
-    if @label != other_obj.label
-      return false
-    end
-
-    return true
-  end
-end
-
-# Dummy parent object for testing: Behave like an array, except for a new method
-#  which returns a DummyObject
-#
-# This doesn't behave exactly like Fog as we save the new object as soon as it is
-# created (not on DummyObject.save) but it will suffice for our needs.
-class DummyParentObject < Array
-  def new(*args)
-    new_obj = DummyObject.new(*args)
-    push(new_obj)
-    return new_obj
-  end
+  module_function :dummy_parent_object
 end
 
 describe 'CMObjBase' do
-
-  #
-  # Sanity tests around our Dummy objects
-  #
-  describe 'Dummy Object Sanity Checks' do
-    describe 'DummyObject' do
-      it 'creates a dummy object properly' do
-        target_obj = DummyObject.new(id: 'id1', label: 'label1')
-        target_obj.id.should eql 'id1'
-        target_obj.label.should eql 'label1'
-        target_obj.saved.should eql false
-        target_obj.destroyed.should eql false
+  describe '#obj_paginated_find' do
+    before :each do
+      @parent_obj = CMObjBaseTestHelpers.dummy_parent_object
+      if @parent_obj != []
+        fail '@parent_obj not empty after initialization'
       end
     end
 
-    describe 'DummyParentObject' do
-      before :all do
-        @parent_obj = DummyParentObject.new
-      end
+    it 'locates all objects in a paginated response' do
+      # Limit to 10 for pagination testing (5 pages)
+      @test_obj = CMObjBase.new(find_pagination_limit: 10)
 
-      it 'is initially an empty array' do
-        @parent_obj.should == []
+      # Seed the mock object with 50 objects
+      50.times do |c|
+        @parent_obj.new(label: "Test Object #{c}").save
       end
+      @parent_obj.length.should eql 50
 
-      it 'allows creation of DummyObjects' do
-        @parent_obj.new(id: 'id1', label: 'label1')
-        @parent_obj[0].id.should eql 'id1'
-        @parent_obj[0].label.should eql 'label1'
-        @parent_obj[0].saved.should eql false
-        @parent_obj[0].destroyed.should eql false
+      # Find them!
+      50.times do |c|
+        @test_obj.obj_paginated_find(@parent_obj, 'paginated find test') { |o| o.label == "Test Object #{c}" }.should eql @parent_obj[c]
       end
+    end
+
+    it 'Passes find_pagination_limit through to the API' do
+      # Test via illegal value
+      @test_obj = CMObjBase.new(find_pagination_limit: 0)
+      expect { @test_obj.obj_paginated_find(@parent_obj, 'find_pagination_limit test') { |o| o.label == 'Test Object 0' } }.to raise_exception
     end
   end
 
-  #
-  # Actual tests
-  #
   describe '#obj_lookup_by_id' do
     before :each do
       @test_obj = CMObjBase.new
-      @parent_obj = DummyParentObject.new
+      @parent_obj = CMObjBaseTestHelpers.dummy_parent_object
       if @parent_obj != []
         fail '@parent_obj not empty after initialization'
       end
@@ -130,20 +76,21 @@ describe 'CMObjBase' do
     end
 
     it 'finds objects in the parent' do
-      target_obj = @parent_obj.new(id: 'id1', label: 'label1')
-      @test_obj.obj_lookup_by_id(nil, @parent_obj, 'obj find test', 'id1').should eql target_obj
+      target_obj = @parent_obj.new(label: 'label1')
+      target_obj.save
+      @test_obj.obj_lookup_by_id(nil, @parent_obj, 'obj find test', target_obj.id).should eql target_obj
     end
 
     it 'returns nil when an object is not found' do
       8.times do |i|
-        @parent_obj.new(id: "id#{i}", label: "label#{i}")
+        @parent_obj.new(label: "label#{i}").save
       end
       @test_obj.obj_lookup_by_id(nil, @parent_obj, 'obj not found test', 'id9').should eql nil
     end
 
     it 'does not search the parent when the target is passed' do
-      target_obj = DummyObject.new(id: 'id1', label: 'label1')
-      # Note target_obj is NOT in the parent array and the parent array is empty
+      target_obj = @parent_obj.new(label: 'label1')
+      # Note target_obj is NOT in the parent array and the parent array is empty as we didn't call save
       # If obj_lookup_by_id fails to return target_obj it should return nil per
       #  earlier test 'returns nil when there are no objects in the parent'
       @test_obj.obj_lookup_by_id(target_obj, @parent_obj, 'existing obj hit test', target_obj.id).should eql target_obj
@@ -153,7 +100,7 @@ describe 'CMObjBase' do
   describe '#obj_lookup_by_label' do
     before :each do
       @test_obj = CMObjBase.new
-      @parent_obj = DummyParentObject.new
+      @parent_obj = CMObjBaseTestHelpers.dummy_parent_object
       if @parent_obj != []
         fail '@parent_obj not empty after initialization'
       end
@@ -164,20 +111,21 @@ describe 'CMObjBase' do
     end
 
     it 'finds objects in the parent' do
-      target_obj = @parent_obj.new(id: 'id1', label: 'label1')
+      target_obj = @parent_obj.new(label: 'label1')
+      target_obj.save
       @test_obj.obj_lookup_by_label(nil, @parent_obj, 'obj find test', 'label1').should eql target_obj
     end
 
     it 'returns nil when an object is not found' do
       8.times do |i|
-        @parent_obj.new(id: "id#{i}", label: "label#{i}")
+        @parent_obj.new(label: "label#{i}").save
       end
       @test_obj.obj_lookup_by_label(nil, @parent_obj, 'obj not found test', 'label9').should eql nil
     end
 
     it 'does not search the parent when the target is passed' do
-      target_obj = DummyObject.new(id: 'id1', label: 'label1')
-      # Note target_obj is NOT in the parent array and the parent array is empty
+      target_obj = @parent_obj.new(label: 'label1')
+      # Note target_obj is NOT in the parent array and the parent array is empty as we didn't call save
       # If obj_lookup_by_label fails to return target_obj it should return nil per
       #  earlier test 'returns nil when there are no objects in the parent'
       @test_obj.obj_lookup_by_label(target_obj, @parent_obj, 'existing obj hit test', target_obj.label).should eql target_obj
@@ -187,48 +135,55 @@ describe 'CMObjBase' do
   describe '#obj_update' do
     before :each do
       @test_obj = CMObjBase.new
-      @parent_obj = DummyParentObject.new
+      @parent_obj = CMObjBaseTestHelpers.dummy_parent_object
     end
 
     it 'creates a new object when obj is nil' do
-      target_obj = @test_obj.obj_update(nil, @parent_obj, 'new obj test', id: 'id1', label: 'label1')
-      target_obj.id.should eql 'id1'
+      target_obj = @test_obj.obj_update(nil, @parent_obj, 'new obj test', label: 'label1')
       target_obj.label.should eql 'label1'
-      target_obj.saved.should eql true
-      target_obj.destroyed.should eql false
+      @parent_obj.length.should eql 1
+      @parent_obj[0].should eql target_obj
     end
 
     it 'replaces an object when the object exists' do
-      orig_target_obj = @parent_obj.new(id: 'id1', label: 'label1')
-      target_obj = @test_obj.obj_update(orig_target_obj, @parent_obj, 'obj replacement test', id: 'id1', label: 'label2')
-      # This test is flawed in that it doesn't test for replacement.
-      # However the current framework doesn't allow for that
-      target_obj.id.should eql 'id1'
+      orig_target_obj = @parent_obj.new(label: 'label1')
+      orig_target_obj.save
+      @parent_obj[0].should eql orig_target_obj
+
+      target_obj = @test_obj.obj_update(orig_target_obj, @parent_obj, 'obj replacement test', label: 'label2')
+
+      target_obj.should_not eql orig_target_obj
+      target_obj.id.should eql orig_target_obj.id
       target_obj.label.should eql 'label2'
-      target_obj.saved.should eql true
-      target_obj.destroyed.should eql false
+
+      @parent_obj.length.should eql 1
+      @parent_obj[0].should eql target_obj
     end
 
     it 'does not modify if not required' do
-      orig_target_obj = @parent_obj.new(id: 'id1', label: 'label1')
-      @test_obj.obj_update(orig_target_obj, @parent_obj, 'obj modification test', id: 'id1', label: 'label1').should eql orig_target_obj
+      orig_target_obj = @parent_obj.new(label: 'label1')
+      @test_obj.obj_update(orig_target_obj, @parent_obj, 'obj modification test', label: 'label1').should eql orig_target_obj
+      @parent_obj.length.should eql 0
     end
   end
 
   describe '#obj_delete' do
     before :each do
       @test_obj = CMObjBase.new
-      @parent_obj = DummyParentObject.new
-      @target_obj = @parent_obj.new(id: 'id1', label: 'label1')
+      @parent_obj = CMObjBaseTestHelpers.dummy_parent_object
+      @target_obj = @parent_obj.new(label: 'label1')
+      @target_obj.save
+      fail 'Failed to save target' if @parent_obj.length != 1
     end
 
     it 'destroys obj' do
       @test_obj.obj_delete(@target_obj, @parent_obj, 'destruction test').should eql true
-      @target_obj.destroyed.should eql true
+      @parent_obj.length.should eql 0
     end
 
     it 'does not try and destroy nil objects' do
       @test_obj.obj_delete(nil, @parent_obj, 'destruction test').should eql false
+      @parent_obj.length.should eql 1
     end
   end
 end