+++ spec/unit/libraries/mock_data_spec.rb
@@ -42,6 +42,7 @@ describe 'mock_data' do
     end
 
     describe '#agent_tokens' do
+      # Using :all instead of :each to allow the destroy test to use the value added by the save test.
       before :all do
         @mock_obj = MockMonitoring.new(rackspace_api_key:  'porkchop',
                                        rackspace_username: 'sandwitches')
@@ -98,9 +99,21 @@ describe 'mock_data' do
       end
 
       describe '#save' do
-        it 'should save the entity into the parent' do
+        it 'should save the token into the parent' do
           @mock_obj.agent_tokens.length.should eql 0
           test_obj = @mock_obj.agent_tokens.new
+          test_obj.save
+          @mock_obj.agent_tokens.length.should eql 1
+          @mock_obj.agent_tokens[0].should eql test_obj
+        end
+
+        it 'should overwrite tokens with duplicate IDs' do
+          @mock_obj.agent_tokens.length.should eql 1
+          test_obj = @mock_obj.agent_tokens[0]
+
+          test_obj.label.should_not eql 'Test label 2'
+          test_obj.label = 'Test label 2'
+
           test_obj.save
           @mock_obj.agent_tokens.length.should eql 1
           @mock_obj.agent_tokens[0].should eql test_obj
@@ -108,7 +121,7 @@ describe 'mock_data' do
       end
 
       describe '#destroy' do
-        it 'should remove the entity from the parent' do
+        it 'should remove the token from the parent' do
           @mock_obj.agent_tokens.length.should eql 1
           target = @mock_obj.agent_tokens[0]
           target.destroy
@@ -129,9 +142,49 @@ describe 'mock_data' do
           test_obj.compare?(test_obj2).should eql false
         end
       end
+
+      describe '#all' do
+        before :each do
+          @mock_obj = MockMonitoring.new(rackspace_api_key:  'porkchop',
+                                         rackspace_username: 'sandwitches')
+          10.times do
+            @mock_obj.agent_tokens.new.save
+          end
+        end
+
+        it 'returns all objects when object count < limit' do
+          all_output = @mock_obj.agent_tokens.all(limit: 20, marker: nil)
+          all_output.should eql @mock_obj.agent_tokens
+          all_output.marker.should eql nil
+        end
+
+        it 'returns all objects when object count == limit' do
+          all_output = @mock_obj.agent_tokens.all(limit: 10, marker: nil)
+          all_output.should eql @mock_obj.agent_tokens
+          all_output.marker.should eql nil
+        end
+
+        it 'Paginates when when object count > limit' do
+          all_output = @mock_obj.agent_tokens.all(limit: 5, marker: nil)
+          all_output.should eql @mock_obj.agent_tokens[0..4]
+          all_output.marker.should eql @mock_obj.agent_tokens[5].id
+
+          all_output = @mock_obj.agent_tokens.all(limit: 5, marker: all_output.marker)
+          all_output.should eql @mock_obj.agent_tokens[5..9]
+          all_output.marker.should eql nil
+        end
+
+        it 'Fails when limit < 1' do
+          expect { @mock_obj.agent_tokens.all(limit: 0, marker: nil) }.to raise_exception
+        end
+        it 'Fails when limit > 1000' do
+          expect { @mock_obj.agent_tokens.all(limit: 1001, marker: nil) }.to raise_exception
+        end
+      end
     end
 
     describe '#entities' do
+      # Using :all instead of :each to allow the destroy test to use the value added by the save test.
       before :all do
         @mock_obj = MockMonitoring.new(rackspace_api_key:  'porkchop',
                                        rackspace_username: 'sandwitches')
@@ -193,6 +246,18 @@ describe 'mock_data' do
           @mock_obj.entities.length.should eql 1
           @mock_obj.entities[0].should eql test_entity
         end
+
+        it 'should overwrite entities with duplicate IDs' do
+          @mock_obj.entities.length.should eql 1
+          test_obj = @mock_obj.entities[0]
+
+          test_obj.label.should_not eql 'Test label 2'
+          test_obj.label = 'Test label 2'
+
+          test_obj.save
+          @mock_obj.entities.length.should eql 1
+          @mock_obj.entities[0].should eql test_obj
+        end
       end
 
       describe '#destroy' do
@@ -215,6 +280,45 @@ describe 'mock_data' do
           test_entity  =  @mock_obj.entities.new
           test_entity2 = @mock_obj.entities.new
           test_entity.compare?(test_entity2).should eql false
+        end
+      end
+
+      describe '#all' do
+        before :each do
+          @mock_obj = MockMonitoring.new(rackspace_api_key:  'porkchop',
+                                         rackspace_username: 'sandwitches')
+          10.times do
+            @mock_obj.entities.new.save
+          end
+        end
+
+        it 'returns all objects when object count < limit' do
+          all_output = @mock_obj.entities.all(limit: 20, marker: nil)
+          all_output.should eql @mock_obj.entities
+          all_output.marker.should eql nil
+        end
+
+        it 'returns all objects when object count == limit' do
+          all_output = @mock_obj.entities.all(limit: 10, marker: nil)
+          all_output.should eql @mock_obj.entities
+          all_output.marker.should eql nil
+        end
+
+        it 'Paginates when when object count > limit' do
+          all_output = @mock_obj.entities.all(limit: 5, marker: nil)
+          all_output.should eql @mock_obj.entities[0..4]
+          all_output.marker.should eql @mock_obj.entities[5].id
+
+          all_output = @mock_obj.entities.all(limit: 5, marker: all_output.marker)
+          all_output.should eql @mock_obj.entities[5..9]
+          all_output.marker.should eql nil
+        end
+
+        it 'Fails when limit < 1' do
+          expect { @mock_obj.entities.all(limit: 0, marker: nil) }.to raise_exception
+        end
+        it 'Fails when limit > 1000' do
+          expect { @mock_obj.entities.all(limit: 1001, marker: nil) }.to raise_exception
         end
       end
 
@@ -272,17 +376,29 @@ describe 'mock_data' do
         end
 
         describe '#save' do
-          it 'should save the entity into the parent' do
+          it 'should save the check into the parent' do
             @test_entity.checks.length.should eql 0
             test_check = @test_entity.checks.new('type' => 'dummy ')
             test_check.save
             @test_entity.checks.length.should eql 1
             @test_entity.checks[0].should eql test_check
           end
+
+          it 'should overwrite checks with duplicate IDs' do
+            @test_entity.checks.length.should eql 1
+            test_obj = @test_entity.checks[0]
+
+            test_obj.label.should_not eql 'Test label 2'
+            test_obj.label = 'Test label 2'
+
+            test_obj.save
+            @test_entity.checks.length.should eql 1
+            @test_entity.checks[0].should eql test_obj
+          end
         end
 
         describe '#destroy' do
-          it 'should remove the entity from the parent' do
+          it 'should remove the check from the parent' do
             @test_entity.checks.length.should eql 1
             target = @test_entity.checks[0]
             target.destroy
@@ -303,9 +419,48 @@ describe 'mock_data' do
             test_check.compare?(test_check2).should eql false
           end
         end
+
+        describe '#all' do
+          before :each do
+            @test_entity = @mock_obj.entities.new
+            10.times do
+              @test_entity.checks.new('type' => 'dummy ').save
+            end
+          end
+
+          it 'returns all objects when object count < limit' do
+            all_output = @test_entity.checks.all(limit: 20, marker: nil)
+            all_output.should eql @test_entity.checks
+            all_output.marker.should eql nil
+          end
+
+          it 'returns all objects when object count == limit' do
+            all_output = @test_entity.checks.all(limit: 10, marker: nil)
+            all_output.should eql @test_entity.checks
+            all_output.marker.should eql nil
+          end
+
+          it 'Paginates when when object count > limit' do
+            all_output = @test_entity.checks.all(limit: 5, marker: nil)
+            all_output.should eql @test_entity.checks[0..4]
+            all_output.marker.should eql @test_entity.checks[5].id
+
+            all_output = @test_entity.checks.all(limit: 5, marker: all_output.marker)
+            all_output.should eql @test_entity.checks[5..9]
+            all_output.marker.should eql nil
+          end
+
+          it 'Fails when limit < 1' do
+            expect { @test_entity.checks.all(limit: 0, marker: nil) }.to raise_exception
+          end
+          it 'Fails when limit > 1000' do
+            expect { @test_entity.checks.all(limit: 1001, marker: nil) }.to raise_exception
+          end
+        end
       end
 
       describe '#alarms' do
+      # Using :all instead of :each to allow the destroy test to use the value added by the save test.
         before :all do
           @test_entity = @mock_obj.entities.new
         end
@@ -363,17 +518,30 @@ describe 'mock_data' do
         end
 
         describe '#save' do
-          it 'should save the entity into the parent' do
+          it 'should save the alarm into the parent' do
             @test_entity.alarms.length.should eql 0
             test_check = @test_entity.alarms.new('check' => 'three', 'notification_plan_id' => 'seven ')
             test_check.save
             @test_entity.alarms.length.should eql 1
             @test_entity.alarms[0].should eql test_check
           end
+
+          it 'should overwrite alarms with duplicate IDs' do
+            @test_entity.alarms.length.should eql 1
+            test_obj = @test_entity.alarms[0]
+
+            test_obj.label.should_not eql 'Test label 2'
+            test_obj.label = 'Test label 2'
+
+            test_obj.save
+            @test_entity.alarms.length.should eql 1
+            @test_entity.alarms[0].should eql test_obj
+          end
+
         end
 
         describe '#destroy' do
-          it 'should remove the entity from the parent' do
+          it 'should remove the alarm from the parent' do
             @test_entity.alarms.length.should eql 1
             target = @test_entity.alarms[0]
             target.destroy
@@ -392,6 +560,44 @@ describe 'mock_data' do
             test_check =  @test_entity.alarms.new('check' => 'three', 'notification_plan_id' => 'seven ')
             test_check2 = @test_entity.alarms.new('check' => 'three', 'notification_plan_id' => 'seven ')
             test_check.compare?(test_check2).should eql false
+          end
+        end
+
+        describe '#all' do
+          before :each do
+            @test_entity = @mock_obj.entities.new
+            10.times do
+              @test_entity.alarms.new('check' => 'three', 'notification_plan_id' => 'seven ').save
+            end
+          end
+
+          it 'returns all objects when object count < limit' do
+            all_output = @test_entity.alarms.all(limit: 20, marker: nil)
+            all_output.should eql @test_entity.alarms
+            all_output.marker.should eql nil
+          end
+
+          it 'returns all objects when object count == limit' do
+            all_output = @test_entity.alarms.all(limit: 10, marker: nil)
+            all_output.should eql @test_entity.alarms
+            all_output.marker.should eql nil
+          end
+
+          it 'Paginates when when object count > limit' do
+            all_output = @test_entity.alarms.all(limit: 5, marker: nil)
+            all_output.should eql @test_entity.alarms[0..4]
+            all_output.marker.should eql @test_entity.alarms[5].id
+
+            all_output = @test_entity.alarms.all(limit: 5, marker: all_output.marker)
+            all_output.should eql @test_entity.alarms[5..9]
+            all_output.marker.should eql nil
+          end
+
+          it 'Fails when limit < 1' do
+            expect { @test_entity.alarms.all(limit: 0, marker: nil) }.to raise_exception
+          end
+          it 'Fails when limit > 1000' do
+            expect { @test_entity.alarms.all(limit: 1001, marker: nil) }.to raise_exception
           end
         end
       end